facialrecognition/requirements.txt
face-recognition==1.3.0
opencv-python==4.5.1.48
numpy==1.19.3
Pillow==8.2.0